refactor : 에러 메시지 테스트 수정
에러 메시지 수정
- 에러 메시지 enum 클래스로 선언(ErrorType)
- 예외 테스트 시, 메시지 비교 로직 추가
- 빈 문자열 예외 처리
