+++ src/test/kotlin/stringcalculator/OperatorTest.kt
@@ -2,7 +2,10 @@ package stringcalculator
 
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.BehaviorSpec
-import java.lang.IllegalArgumentException
+import io.kotest.matchers.should
+import io.kotest.matchers.string.startWith
+import stringcalculator.constants.ErrorType.INVALID_OPERATOR
+import stringcalculator.enums.Operator
 
 class OperatorTest : BehaviorSpec({
     Given("잘못된 연산자를 제공한다.") {
@@ -10,7 +13,8 @@ class OperatorTest : BehaviorSpec({
 
         When("잘못된 연산자로 Operator 를 찾을 때") {
             Then("IllegalArgumentException 예외를 던진다.") {
-                shouldThrow<IllegalArgumentException> { Operator.of(input) }
+                val exception = shouldThrow<IllegalArgumentException> { Operator.of(input) }
+                exception.message should startWith(INVALID_OPERATOR.message)
             }
         }
     }
